Move throw_libsql_error() to separate file
